Migrate benchmark to a crossbuild as well and get JVM working again
