+++ build.sbt
@@ -30,16 +30,25 @@ lazy val xoroshiro128JS     = xoroshiro128.js
 lazy val xoroshiro128Native = xoroshiro128.native
 
 lazy val benchmark =
-  (project in file("benchmark"))
+  crossProject(JVMPlatform, JSPlatform, NativePlatform)
+    .crossType(CrossType.Full)
     .settings(
       commonSettings,
       name := "xoroshiro128-benchmark",
-      mainClass in Compile := Some("Benchmark"),
+      mainClass in Compile := Some("Benchmark")
+    )
+    .jvmSettings(
+      mainClass in Compile := Some("JvmBenchmark"),
 
       // Scalameter for speed testing.
       resolvers += "Sonatype OSS Snapshots" at "https://oss.sonatype.org/content/repositories/releases",
       libraryDependencies += "com.storm-enroute" %% "scalameter-core" % "0.7"
-    ).dependsOn(xoroshiro128JVM)
+    )
+    .dependsOn(xoroshiro128)
+
+lazy val benchmarkJVM    = benchmark.jvm
+lazy val benchmarkJS     = benchmark.js
+lazy val benchmarkNative = benchmark.native
 
 lazy val `scala-xoroshiro128` =
   (project in file("."))
